Move OrdersTable into it's own component
